+++ src/blender_executable.ts
@@ -19,9 +19,9 @@ import { deduplicateSameHardLinks } from './blender_executable_linux';
 const stat = util.promisify(fs.stat)
 
 export class BlenderExecutable {
-    data: BlenderPathData;
+    data: BlenderExecutableData;
 
-    constructor(data: BlenderPathData) {
+    constructor(data: BlenderExecutableData) {
         this.data = data;
     }
 
@@ -92,7 +92,7 @@ export class BlenderExecutable {
             request: 'launch',
             program: this.data.path,
             args: ['--debug'].concat(getBlenderLaunchArgs(blend_filepath)),
-            environment: Object.entries(env).map(([key, value]) => { return {name: key, value}; }),
+            environment: Object.entries(env).map(([key, value]) => { return { name: key, value }; }),
             stopAtEntry: false,
             MIMode: 'gdb',
             cwd: folder.uri.fsPath,
@@ -110,7 +110,14 @@ export class BlenderExecutable {
     }
 }
 
-export interface BlenderPathData {
+export type BlenderExecutableSettings = {
+    path: string;
+    name: string;
+    isDebug: boolean;
+    linuxInode?: never;
+}
+
+export type BlenderExecutableData = {
     path: string;
     name: string;
     isDebug: boolean;
@@ -120,12 +127,12 @@ export interface BlenderPathData {
 interface BlenderType {
     label: string;
     selectNewLabel: string;
-    predicate: (item: BlenderPathData) => boolean;
-    setSettings: (item: BlenderPathData) => void;
+    predicate: (item: BlenderExecutableData) => boolean;
+    setSettings: (item: BlenderExecutableData) => void;
 }
 
-async function searchBlenderInSystem(): Promise<BlenderPathData[]> {
-    const blenders: BlenderPathData[] = [];
+async function searchBlenderInSystem(): Promise<BlenderExecutableData[]> {
+    const blenders: BlenderExecutableData[] = [];
     if (process.platform === "win32") {
         const windowsBlenders = await getBlenderWindows();
         blenders.push(...windowsBlenders.map(blend_path => ({ path: blend_path, name: "", isDebug: false })))
@@ -145,11 +152,11 @@ async function searchBlenderInSystem(): Promise<BlenderPathData[]> {
     return blenders;
 }
 
-async function getFilteredBlenderPath(type: BlenderType): Promise<BlenderPathData> {
-    let result: BlenderPathData[] = []
+async function getFilteredBlenderPath(type: BlenderType): Promise<BlenderExecutableData> {
+    let result: BlenderExecutableData[] = []
     {
-        const blenderPathsInSystem: BlenderPathData[] = await searchBlenderInSystem();
-        const deduplicatedBlenderPaths: BlenderPathData[] = deduplicateSamePaths(blenderPathsInSystem);
+        const blenderPathsInSystem: BlenderExecutableData[] = await searchBlenderInSystem();
+        const deduplicatedBlenderPaths: BlenderExecutableData[] = deduplicateSamePaths(blenderPathsInSystem);
         if (process.platform !== 'win32') {
             try {
                 result = await deduplicateSameHardLinks(deduplicatedBlenderPaths, true);
@@ -162,9 +169,9 @@ async function getFilteredBlenderPath(type: BlenderType): Promise<BlenderPathDat
     }
 
     const config = getConfig();
-    const settingsBlenderPaths = (<BlenderPathData[]>config.get('executables')).filter(type.predicate);
+    const settingsBlenderPaths = (<BlenderExecutableData[]>config.get('executables')).filter(type.predicate);
     { // deduplicate Blender paths twice: it preserves proper order in UI
-        const deduplicatedBlenderPaths: BlenderPathData[] = deduplicateSamePaths(result, settingsBlenderPaths);
+        const deduplicatedBlenderPaths: BlenderExecutableData[] = deduplicateSamePaths(result, settingsBlenderPaths);
         if (process.platform !== 'win32') {
             try {
                 result = [...settingsBlenderPaths, ...await deduplicateSameHardLinks(deduplicatedBlenderPaths, false, settingsBlenderPaths)]
@@ -189,19 +196,20 @@ async function getFilteredBlenderPath(type: BlenderType): Promise<BlenderPathDat
     quickPickItems.push({ label: type.selectNewLabel, data: async () => askUser_FilteredBlenderPath(type) })
 
     const pickedItem = await letUserPickItem(quickPickItems);
-    const pathData: BlenderPathData = await pickedItem.data();
+    const pathData: BlenderExecutableData = await pickedItem.data();
 
     // update VScode settings
     if (settingsBlenderPaths.find(data => data.path === pathData.path) === undefined) {
         settingsBlenderPaths.push(pathData);
-        config.update('executables', settingsBlenderPaths, vscode.ConfigurationTarget.Global);
+        const toSave: BlenderExecutableSettings[] = settingsBlenderPaths.map(item => { return { 'name': item.name, 'isDebug': item.isDebug, 'path': item.path } })
+        config.update('executables', toSave, vscode.ConfigurationTarget.Global);
     }
 
     return pathData;
 }
 
-function deduplicateSamePaths(blenderPathsToReduce: BlenderPathData[], additionalBlenderPaths: BlenderPathData[] = []) {
-    const deduplicatedBlenderPaths: BlenderPathData[] = [];
+function deduplicateSamePaths(blenderPathsToReduce: BlenderExecutableData[], additionalBlenderPaths: BlenderExecutableData[] = []) {
+    const deduplicatedBlenderPaths: BlenderExecutableData[] = [];
     const uniqueBlenderPaths: string[] = [];
     const isTheSamePath = (path_one: string, path_two: string) => path.relative(path_one, path_two) === '';
     for (const item of blenderPathsToReduce) {
@@ -217,9 +225,9 @@ function deduplicateSamePaths(blenderPathsToReduce: BlenderPathData[], additiona
     return deduplicatedBlenderPaths;
 }
 
-async function askUser_FilteredBlenderPath(type: BlenderType): Promise<BlenderPathData> {
+async function askUser_FilteredBlenderPath(type: BlenderType): Promise<BlenderExecutableData> {
     let filepath = await askUser_BlenderPath(type.label);
-    let pathData: BlenderPathData = {
+    let pathData: BlenderExecutableData = {
         path: filepath,
         name: '',
         isDebug: false,